requirements.txt
google-genai 
python-dotenv
Django==5.2.3
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
binaryornot==0.4.4
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
executing==2.2.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
narwhals==2.3.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.101.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
pip==25.2
platformdirs==4.4.0
plyer==2.1.0
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
openai
django-allauth
LLMAudit
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.19
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
